requirements.txt
torch==1.7.1
torchvision
matplotlib
tensorboardX
numpy
librosa==0.7.2
inflect==0.2.5
scipy
Unidecode==1.0.22
pillow
tqdm
configargparse
